# docs/api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "DomusUnify API",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/item-categories": {
      "get": {
        "tags": [
          "ItemCategories"
        ],
        "summary": "Obtém todas as categorias de itens para a família atual.",
        "description": "Lista todas as categorias visíveis para o utilizador na família atual.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ItemCategories"
        ],
        "summary": "Cria uma nova categoria de itens para a família atual.",
        "description": "Permite definir nome, ícone e ordem de classificação. O utilizador deve ter permissões na família.",
        "requestBody": {
          "description": "O pedido contendo detalhes da categoria a criar.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/item-categories/{categoryId}": {
      "patch": {
        "tags": [
          "ItemCategories"
        ],
        "summary": "Atualiza uma categoria de itens existente.",
        "description": "Permite alterar nome/ícone e reordenar a categoria via `SortOrder`.",
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "description": "O ID da categoria a atualizar.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "O pedido contendo detalhes atualizados da categoria.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ItemCategories"
        ],
        "summary": "Elimina uma categoria de itens.",
        "description": "Remove a categoria se não estiver em uso. O utilizador deve ter permissões na família.",
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "description": "O ID da categoria a eliminar.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/lists/items/{itemId}": {
      "patch": {
        "tags": [
          "Lists"
        ],
        "summary": "Atualiza um item de lista existente.",
        "description": "Permite alterar nome, estado de conclusão e categoria. CategoryId pode ser null para remover categoria.",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "description": "O ID do item a atualizar.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "O pedido contendo detalhes atualizados do item.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateListItemRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateListItemRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateListItemRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Lists"
        ],
        "summary": "Elimina um item de lista existente.",
        "description": "Remove o item da lista.",
        "parameters": [
          {
            "name": "itemId",
            "in": "path",
            "description": "O ID do item a eliminar.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CategoryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identificador da categoria.",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Nome da categoria.",
            "nullable": true
          },
          "iconKey": {
            "type": "string",
            "description": "Chave do ícone da categoria.",
            "nullable": true
          },
          "sortOrder": {
            "type": "integer",
            "description": "Ordem de apresentação.",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "Resposta com informação de uma categoria de itens (listas)."
      },
...
      "CreateCategoryRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nome da categoria.",
            "nullable": true
          },
          "iconKey": {
            "type": "string",
            "description": "Chave do ícone da categoria.",
            "nullable": true
          },
          "sortOrder": {
            "type": "integer",
            "description": "Ordem de apresentação (opcional).",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "Pedido para criar uma nova categoria de itens (listas)."
      },
...
      "UpdateCategoryRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Novo nome da categoria (opcional).",
            "nullable": true
          },
          "iconKey": {
            "type": "string",
            "description": "Nova chave do ícone (opcional).",
            "nullable": true
          },
          "sortOrder": {
            "type": "integer",
            "description": "Nova ordem de apresentação (opcional).",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Pedido para atualizar uma categoria de itens (listas)."
      },
...
      "UpdateListItemRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Novo nome do item (opcional).",
            "nullable": true
          },
          "isCompleted": {
            "type": "boolean",
            "description": "Indica se o item passa a estar concluído (opcional).",
            "nullable": true
          },
          "categoryId": {
            "description": "Identificador da categoria do item (opcional).",
            "nullable": true
          },
          "assigneeUserId": {
            "description": "Identificador do utilizador a quem o item estÃ¡ atribuÃ­do (opcional).",
            "nullable": true
          },
          "note": {
            "description": "Nota do item (opcional).",
            "nullable": true
          },
          "photoUrl": {
            "description": "URL (ou data URL) da foto do item (opcional).",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Pedido para atualizar um item de lista."
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Insere o token JWT assim: **Bearer {token}**",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}